requirements.txt
disnake
imageio
numpy
pillow
logging